requirements.txt
streamlit
opencv-python-headless
pytesseract
pillow
google-genai
python-dotenv